refactor(utils): reorganize browser exports and move Effect-based browser utils
Move browser-specific Effect utilities (Opfs, WebError, WebLock, QuotaExceededError)
from effect/ to browser/ directory and update all imports across the codebase.
