+++ packages/@livestore/adapter-cloudflare/src/make-adapter.ts
@@ -12,8 +12,7 @@ import type { CfTypes } from '@livestore/common-cf'
 import { LiveStoreEvent } from '@livestore/livestore'
 import { sqliteDbFactory } from '@livestore/sqlite-wasm/cf'
 import { loadSqlite3Wasm } from '@livestore/sqlite-wasm/load-wasm'
-import { Effect, FetchHttpClient, Layer, Schedule, SubscriptionRef } from '@livestore/utils/effect'
-import { WebChannel } from '@livestore/utils/effect/browser'
+import { Effect, FetchHttpClient, Layer, Schedule, SubscriptionRef, WebChannel } from '@livestore/utils/effect'
 
 export const makeAdapter =
   ({
