+++ packages/@livestore/adapter-node/src/webchannel.ts
@@ -1,8 +1,7 @@
 import type { BroadcastChannel as NodeBroadcastChannel } from 'node:worker_threads'
 
 import type { Either, ParseResult } from '@livestore/utils/effect'
-import { Deferred, Effect, Exit, Schema, Scope, Stream } from '@livestore/utils/effect'
-import { WebChannel } from '@livestore/utils/effect/browser'
+import { Deferred, Effect, Exit, Schema, Scope, Stream, WebChannel } from '@livestore/utils/effect'
 
 export const makeBroadcastChannel = <Msg, MsgEncoded>({
   channelName,
