+++ packages/@livestore/adapter-web/src/in-memory/in-memory-adapter.ts
@@ -18,16 +18,8 @@ import type { MakeWebSqliteDb } from '@livestore/sqlite-wasm/browser'
 import { sqliteDbFactory } from '@livestore/sqlite-wasm/browser'
 import { tryAsFunctionAndNew } from '@livestore/utils'
 import type { Scope } from '@livestore/utils/effect'
-import {
-  BrowserWorker,
-  Effect,
-  FetchHttpClient,
-  Fiber,
-  Layer,
-  type Schema,
-  SubscriptionRef,
-  Worker,
-} from '@livestore/utils/effect'
+import { Effect, FetchHttpClient, Fiber, Layer, type Schema, SubscriptionRef, Worker } from '@livestore/utils/effect'
+import { BrowserWorker } from '@livestore/utils/effect/browser'
 import { nanoid } from '@livestore/utils/nanoid'
 import * as Webmesh from '@livestore/webmesh'
