+++ packages/@livestore/adapter-web/src/web-worker/common/shutdown-channel.ts
@@ -1,8 +1,15 @@
 import { ShutdownChannel } from '@livestore/common/leader-thread'
-import { WebChannel } from '@livestore/utils/effect/browser'
+import type { Effect, Scope, WebChannel } from '@livestore/utils/effect'
+import { WebChannelBrowser } from '@livestore/utils/effect/browser'
 
-export const makeShutdownChannel = (storeId: string) =>
-  WebChannel.broadcastChannel({
+export const makeShutdownChannel = (
+  storeId: string,
+): Effect.Effect<
+  WebChannel.WebChannel<typeof ShutdownChannel.All.Type, typeof ShutdownChannel.All.Type>,
+  never,
+  Scope.Scope
+> =>
+  WebChannelBrowser.broadcastChannel({
     channelName: `livestore.shutdown.${storeId}`,
     schema: ShutdownChannel.All,
   })
