+++ packages/@livestore/common/src/devtools/mod.ts
@@ -1,6 +1,5 @@
-import type { Effect, Scope } from '@livestore/utils/effect'
+import type { Effect, Scope, WebChannel } from '@livestore/utils/effect'
 import { Schema } from '@livestore/utils/effect'
-import type { WebChannel } from '@livestore/utils/effect/browser'
 import { nanoid } from '@livestore/utils/nanoid'
 import type { MeshNode } from '@livestore/webmesh'
