+++ packages/@livestore/devtools-web-common/src/worker/mod.ts
@@ -1,6 +1,5 @@
 import { LS_DEV } from '@livestore/utils'
-import { Context, Effect, Layer, Stream } from '@livestore/utils/effect'
-import { WebChannel } from '@livestore/utils/effect/browser'
+import { Context, Effect, Layer, Stream, WebChannel } from '@livestore/utils/effect'
 import type { MeshNode } from '@livestore/webmesh'
 import { makeMeshNode, WebmeshSchema } from '@livestore/webmesh'
