+++ packages/@livestore/livestore/src/live-queries/base-class.ts
@@ -23,14 +23,15 @@ export type ReactivityGraphContext = {
   effectsWrapper: (run: () => void) => void
 }
 
-export type GetResult<TQuery extends LiveQueryDef.Any | LiveQuery.Any | SignalDef<any>> =
-  TQuery extends LiveQuery<infer TResult>
+export type GetResult<TQuery extends LiveQueryDef.Any | LiveQuery.Any | SignalDef<any>> = TQuery extends LiveQuery<
+  infer TResult
+>
+  ? TResult
+  : TQuery extends LiveQueryDef<infer TResult>
     ? TResult
-    : TQuery extends LiveQueryDef<infer TResult>
+    : TQuery extends SignalDef<infer TResult>
       ? TResult
-      : TQuery extends SignalDef<infer TResult>
-        ? TResult
-        : unknown
+      : unknown
 
 let queryIdCounter = 0
