+++ packages/@livestore/utils/src/browser/QuotaExceededError.ts
