+++ packages/@livestore/utils/src/browser/WebError.test.ts
