+++ packages/@livestore/utils/src/browser/WebError.ts
