--- packages/@livestore/utils/src/browser/effect.ts
@@ -1,8 +1,0 @@
-import '../global.ts'
-
-export { BrowserWorker, BrowserWorkerRunner } from '@effect/platform-browser'
-export * as Opfs from '../effect/Opfs/mod.ts'
-export * as WebChannel from '../effect/WebChannel/mod.ts'
-export * as WebError from '../effect/WebError.ts'
-export * as WebLock from '../effect/WebLock.ts'
-export * as WebSocket from '../effect/WebSocket.ts'
