+++ packages/@livestore/utils/src/effect/mod.ts
@@ -154,3 +154,5 @@ export * as Stream from './Stream.ts'
 export * as Subscribable from './Subscribable.ts'
 export * as SubscriptionRef from './SubscriptionRef.ts'
 export * as TaskTracing from './TaskTracing.ts'
+export * as WebChannel from './WebChannel/mod.ts'
+export * as WebSocket from './WebSocket.ts'
