+++ packages/@livestore/webmesh/src/node.test.ts
@@ -1,6 +1,5 @@
 import { IS_CI, omitUndefineds } from '@livestore/utils'
-import { Chunk, Deferred, Effect, Exit, Schema, Scope, Stream } from '@livestore/utils/effect'
-import { WebChannel } from '@livestore/utils/effect/browser'
+import { Chunk, Deferred, Effect, Exit, Schema, Scope, Stream, WebChannel } from '@livestore/utils/effect'
 import { Vitest } from '@livestore/utils-dev/node-vitest'
 import { expect } from 'vitest'
