+++ tests/integration/src/tests/playwright/unit-tests/schema-migration/index.ts
@@ -2,7 +2,8 @@ import { makePersistedAdapter } from '@livestore/adapter-web'
 import LiveStoreSharedWorker from '@livestore/adapter-web/shared-worker?sharedworker'
 import type { BootStatus } from '@livestore/common'
 import { liveStoreStorageFormatVersion, UnknownError } from '@livestore/common'
-import { Chunk, Effect, Layer, Logger, LogLevel, Opfs, Queue, Schedule, Schema, Stream } from '@livestore/utils/effect'
+import { Chunk, Effect, Layer, Logger, LogLevel, Queue, Schedule, Schema, Stream } from '@livestore/utils/effect'
+import { Opfs } from '@livestore/utils/effect/browser'
 import { ResultMultipleMigrations } from '../bridge.ts'
 import LiveStoreWorker from '../livestore.worker.ts?worker'
 import { schema } from '../schema.ts'
